requirements.txt
attrs==21.4.0
autopep8==1.6.0
cycler==0.11.0
flake8==4.0.1
iniconfig==1.1.1
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.19.5
packaging==21.3
Pillow==8.3.1
pkg_resources==0.0.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.10.0+cu111
torchvision==0.11.1+cu111
typing_extensions==4.1.1
